better plane spawning
plane spawns now use one of two gameobjects/colliders in scene as an area to spawn in, depending on if enemy or friendly. planes will also be be flipped and fly the other way, as will their torpedoes, if from the enemy.
airstrikelaunchinfo used here so not to disrupt the weapon init methods