aa 2
only been working on this off and on recently.
fleet's aa is controlled by a single fleetaacontroller that is attached to the lead vanguard ship. a list of the aa gun objects (still attached to their ships, even though they are basically dummy objects) is given to the controller. currently using a gizmo to show the aa range (which with double radius seems only slightly larger than it should be), the controller prints a message to show when it is firing, and only does so when planes are in range. no damage/firing is actually done yet though